[refactor] ControllerTransitionState: extract enum into separate class and add unit test
