+++ app/src/main/java/com/rwmobi/dazncodechallenge/ui/components/PictureInPictureButton.kt
@@ -10,13 +10,18 @@ package com.rwmobi.dazncodechallenge.ui.components
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import com.rwmobi.dazncodechallenge.R
 
 @Composable
-fun PictureInPictureButton(onClick: () -> Unit) {
+fun PictureInPictureButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit,
+) {
     IconButton(
+        modifier = modifier,
         onClick = onClick,
     ) {
         Icon(
